recursion: Stack
        sort a stack
        reverse a stack (space complexity: O(1))

+++ basics/recursion/practice/sortAStack.cpp
@@ -7,7 +7,8 @@ public:
     stack<int> s;
 
 void sort() {
-     if(s.size() <= 1) {
+     if(s.size() <= static_cast<size_t>(1)) {
+     //if(s.size() <= 1) {
          return;
      }
      int topVal = s.top();
